requirements.txt
streamlit>=1.30.0
streamlit-sortables>=0.2.0
openai>=1.3.0
reportlab>=4.0.7
python-dotenv>=1.0.0
Pillow>=10.0.0
